Changed a bunch of debug and information printfs so that they have a better chance of working on systems where int is smaller than 32 bits also fixed the signed/unsigned type of some to match the type of the arguments.

+++ bacnet-stack/demo/handler/h_arf_a.c
@@ -79,8 +79,8 @@ void handler_atomic_read_file_ack(
                             octetstring_length(&data.fileData), 1,
                             pFile) != 1) {
 #if PRINT_ENABLED
-                        fprintf(stderr, "Failed to write to %s (%u)!\n",
-                            pFilename, instance);
+                        fprintf(stderr, "Failed to write to %s (%lu)!\n",
+                            pFilename, (unsigned long)instance);
 #endif
                     }
                     fclose(pFile);
