+++ bacnet-stack/demo/object/av.c
@@ -215,7 +215,7 @@ char *Analog_Value_Name(
     static char text_string[32] = "";   /* okay for single thread */
 
     if (object_instance < MAX_ANALOG_VALUES) {
-        sprintf(text_string, "ANALOG VALUE %u", object_instance);
+        sprintf(text_string, "ANALOG VALUE %lu", (unsigned long)object_instance);
         return text_string;
     }
